- Moved both program and Test population in folder. - added more tests for unit test - Renamed UnitedTest => TestPopulation - No longer writes in folder when Formating exaption is caught, instead provides aproperiate message
