+++ Population/Population.sln
@@ -3,7 +3,7 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
 VisualStudioVersion = 17.5.002.0
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ReservationSystem", "ReservationSystem\ReservationSystem.csproj", "{620CBE60-BF8F-442C-B783-66B02098CEF1}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Population", "Population.csproj", "{6B9708A0-5431-472C-9B8E-4837EA316D95}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -11,15 +11,15 @@ Global
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{620CBE60-BF8F-442C-B783-66B02098CEF1}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{620CBE60-BF8F-442C-B783-66B02098CEF1}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{620CBE60-BF8F-442C-B783-66B02098CEF1}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{620CBE60-BF8F-442C-B783-66B02098CEF1}.Release|Any CPU.Build.0 = Release|Any CPU
+		{6B9708A0-5431-472C-9B8E-4837EA316D95}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{6B9708A0-5431-472C-9B8E-4837EA316D95}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{6B9708A0-5431-472C-9B8E-4837EA316D95}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{6B9708A0-5431-472C-9B8E-4837EA316D95}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {4987FC4F-2838-4A3E-89B6-6BB892B480AB}
+		SolutionGuid = {E634009C-F844-41BE-B099-0CE84421C22E}
 	EndGlobalSection
 EndGlobal
