+++ Population/Program.cs
@@ -1,0 +1,8 @@
+﻿public static class Program
+{
+    public static void Main()
+    {
+        Population pop = new Population();
+        pop.AddVistors();
+    }
+}
