--- Program.cs
@@ -1,8 +1,0 @@
-﻿public static class Program
-{
-    public static void Main()
-    {
-        Population pop = new Population();
-        pop.AddVistors();
-    }
-}
