+++ TestPopulation/TestPopulation.cs
@@ -1,39 +1,42 @@
-namespace Population;
+using Microsoft.VisualStudio.TestPlatform.Common.Utilities;
+using Newtonsoft.Json;
+
+namespace TestPopulation;
 
 [TestClass]
-public class TestPopulation;
-
+public class UnitTest1
 {
     [TestMethod]
-    public void TestFileCreation()
+    public void FileCreation()
     {
         bool FileExist = false;
         Population pop = new Population();
         DateTime currentDate = DateTime.Today;
         string filePath = $"Population_on_day_{currentDate:dd-MM-yyyy}.json";
         pop.JSONWriter(0, filePath, currentDate);  
-        Assert.IsTrue(File.Exists(filePath));  // Check if file is made
+        Assert.IsTrue(File.Exists(filePath));  // Check if (new) file is made
     }
+    [TestMethod]
     public void TestTrackVistorOnebyOne()
     {
         Population pop = new Population();
         DateTime currentDate = DateTime.Today;
         string filePath = $"Population_on_day_{currentDate:dd-MM-yyyy}.json";
-        pop.AddVistors(1) // add one visitor to JSON (+1)
-        pop.AddVistors(1) // add one visitor to JSON (+1)
-        pop.AddVistors(1) // add one visitor to JSON (+1)
+        pop.AddVistors(1); // add one visitor to JSON (+1)
+        pop.AddVistors(1); // add one visitor to JSON (+1)
+        pop.AddVistors(1); // add one visitor to JSON (+1)
         // make read JSON method to get an int from JSON
-        Result = Totalvisitor % 3
-        Assert.IsEqual(Result, 0)
+        int Result = pop.TotalVisitor % 3;  // logic here is that it gets and writes to JSON note that this test will fail if you us
+        Assert.AreEqual(Result, 0);
     }
+    [TestMethod]
     public void TestTrackVistorAllinOne()
     {
         Population pop = new Population();
         DateTime currentDate = DateTime.Today;
         string filePath = $"Population_on_day_{currentDate:dd-MM-yyyy}.json";
-        pop.AddVistors(3) // add three visitors to JSON (+3)
-        // make read JSON method to get an int from JSON
-        Result = Totalvisitor % 3
-        Assert.IsEqual(Result, 0)
+        pop.AddVistors(3); // add three visitors to JSON (+3)
+        int Result = pop.TotalVisitor % 3;  
+        Assert.AreEqual(Result, 0);
     }
 }
